fix: character ID overflow, missing trial_daily_stats table, avatar not saved
- Change (c->>'id')::int to ::bigint in character thumbnails query (IDs like
  1773077392247 overflow PostgreSQL integer max)
- Add trial_daily_stats CREATE TABLE to server.js initializeDatabase (was only
  in unused database.js copy, causing repeated errors)
- Create placeholder character entry when photo re-upload can't find character
  in DB (wizard hasn't saved yet), so avatar job can find it later instead of
  retrying 30 times and discarding the generated avatar

+++ server/routes/characters.js
@@ -150,7 +150,7 @@ router.get('/thumbnails', authenticateToken, async (req, res) => {
     const result = await dbQuery(`
       SELECT jsonb_agg(
         jsonb_build_object(
-          'id', (c->>'id')::int,
+          'id', (c->>'id')::bigint,
           'faceThumbnails', c->'avatars'->'faceThumbnails',
           'bodyThumbnails', c->'avatars'->'bodyThumbnails'
         )
@@ -164,7 +164,7 @@ router.get('/thumbnails', authenticateToken, async (req, res) => {
       const legacy = await dbQuery(`
         SELECT jsonb_agg(
           jsonb_build_object(
-            'id', (c->>'id')::int,
+            'id', (c->>'id')::bigint,
             'faceThumbnails', c->'avatars'->'faceThumbnails',
             'bodyThumbnails', c->'avatars'->'bodyThumbnails'
           )
